Alle mine noter fra Excel-arket er lavet. + tilkoblet RankedIn-ID

+++ src/pages/(logged-in)/matchFinder/ViewMatchPage.tsx
@@ -204,7 +204,6 @@ export const ViewMatchPage = () => {
         setError("Invalid match data returned");
         alert("Der opstod en fejl – prøv igen.");
       }
-      alert("Tilmelding sendt!");
       setMatch(updatedMatch);
     } catch (error: any) {
       console.error("Error joining match:", error);
@@ -225,7 +224,6 @@ export const ViewMatchPage = () => {
         setError("Invalid match data returned");
         alert("Der opstod en fejl – prøv igen.");
       }
-      alert("Tilmelding accepteret!");
       setMatch(updatedMatch);
     } catch (error: any) {
       console.error("Error confirming join:", error);
@@ -246,7 +244,6 @@ export const ViewMatchPage = () => {
         setError("Invalid match data returned");
         alert("Der opstod en fejl – prøv igen.");
       }
-      alert("Tilmelding afvist!");
       setMatch(updatedMatch);
     } catch (error: any) {
       console.error("Error confirming join:", error);
@@ -267,7 +264,6 @@ export const ViewMatchPage = () => {
         setError("Invalid match data returned");
         alert("Der opstod en fejl – prøv igen.");
       }
-      alert("Tilmelding bekræftet!");
       setMatch(updatedMatch);
     } catch (error: any) {
       console.error("Error confirming join:", error);
@@ -288,7 +284,6 @@ export const ViewMatchPage = () => {
         setError("Invalid match data returned");
         alert("Der opstod en fejl – prøv igen.");
       }
-      alert("Tilmelding afvist!");
       setMatch(updatedMatch);
     } catch (error: any) {
       console.error("Error confirming join:", error);
@@ -308,7 +303,6 @@ export const ViewMatchPage = () => {
         setError("Invalid match data returned");
         alert("Der opstod en fejl – prøv igen.");
       }
-      alert("Tilmelding annulleret!");
       setMatch(updatedMatch);
     } catch (error: any) {
       console.error("Error confirming join:", error);
@@ -326,7 +320,6 @@ export const ViewMatchPage = () => {
         setError("Invalid match data returned");
         alert("Der opstod en fejl – prøv igen.");
       }
-      alert("Tilmelding afvist!");
       setMatch(updatedMatch);
     } catch (error: any) {
       console.error("Error confirming join:", error);
@@ -338,9 +331,11 @@ export const ViewMatchPage = () => {
   const handleDeleteMatch = async () => {
     if (!match) return;
     try {
+      const userConfirm = confirm("Er du sikker på at du vil slette kampen?");
+      if (userConfirm) {
       await communityApi.deleteMatch(match.id);
-      alert("Kamp slettet!");
       window.history.back();
+      }
     } catch (error: any) {
       console.error("Error deleting match:", error);
       alert(error.response?.data?.message || "Fejl ved sletning af kamp");
